// web/frontend/app/utils/bond/generated/types/terms.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getI64Decoder,
  getI64Encoder,
  getStructDecoder,
  getStructEncoder,
  getU16Decoder,
  getU16Encoder,
  getU64Decoder,
  getU64Encoder,
  getU8Decoder,
  getU8Encoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

/** Условия выпуска: суммы в минимальных единицах платёжного токена, время в секундах unix */
export type Terms = {
  face: bigint;
  couponRateBps: number;
  paymentsPerYear: number;
  nEvents: number;
  startTs: bigint;
  periodSecs: bigint;
  /** Дата выплаты минус дата фиксации */
  recordOffsetSecs: bigint;
  /** Минимум между объявлением погашения и датой фиксации */
  noticeSecs: bigint;
  /** 0: Attestor, 1: Trustless, как `TrustLevel` реестра Mör */
  minTrustLevel: number;
};

export type TermsArgs = {
  face: number | bigint;
  couponRateBps: number;
  paymentsPerYear: number;
  nEvents: number;
  startTs: number | bigint;
  periodSecs: number | bigint;
  /** Дата выплаты минус дата фиксации */
  recordOffsetSecs: number | bigint;
  /** Минимум между объявлением погашения и датой фиксации */
  noticeSecs: number | bigint;
  /** 0: Attestor, 1: Trustless, как `TrustLevel` реестра Mör */
  minTrustLevel: number;
};

export function getTermsEncoder(): FixedSizeEncoder<TermsArgs> {
  return getStructEncoder([
    ["face", getU64Encoder()],
    ["couponRateBps", getU16Encoder()],
    ["paymentsPerYear", getU8Encoder()],
    ["nEvents", getU8Encoder()],
    ["startTs", getI64Encoder()],
    ["periodSecs", getI64Encoder()],
    ["recordOffsetSecs", getI64Encoder()],
    ["noticeSecs", getI64Encoder()],
    ["minTrustLevel", getU8Encoder()],
  ]);
}

export function getTermsDecoder(): FixedSizeDecoder<Terms> {
  return getStructDecoder([
    ["face", getU64Decoder()],
    ["couponRateBps", getU16Decoder()],
    ["paymentsPerYear", getU8Decoder()],
    ["nEvents", getU8Decoder()],
    ["startTs", getI64Decoder()],
    ["periodSecs", getI64Decoder()],
    ["recordOffsetSecs", getI64Decoder()],
    ["noticeSecs", getI64Decoder()],
    ["minTrustLevel", getU8Decoder()],
  ]);
}

export function getTermsCodec(): FixedSizeCodec<TermsArgs, Terms> {
  return combineCodec(getTermsEncoder(), getTermsDecoder());
}